admin/transform/delete: path decode names

+++ src/v/redpanda/admin/transform.cc
@@ -51,7 +51,11 @@ admin_server::delete_transform(std::unique_ptr<ss::http::request> req) {
     if (!_transform_service->local_is_initialized()) {
         throw transforms_not_enabled();
     }
-    auto name = model::transform_name(req->param["name"]);
+    ss::sstring raw_name;
+    if (!admin::path_decode(req->param["name"], raw_name)) {
+        throw seastar::httpd::bad_request_exception("invalid transform name");
+    }
+    auto name = model::transform_name(std::move(raw_name));
     std::error_code ec = co_await _transform_service->local().delete_transform(
       std::move(name));
     co_await throw_on_error(*req, ec, model::controller_ntp);
